requirements.txt
# 核心依赖
requests==2.31.0
python-dotenv==1.0.0

# 开发工具
pytest==7.4.0
flake8==6.0.0
black==23.9.0
